Backend/requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Deep Learning and AI
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
scikit-learn==1.3.0
xgboost==1.7.6

# Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
nibabel==5.1.0
SimpleITK==2.3.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3

# PDF Generation and Reporting
reportlab==4.0.4

# Web and API
requests==2.31.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Medical Imaging (Optional)
pydicom==2.4.3

# Development and Testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Security
cryptography==41.0.4